Inline signal view query hooks
